requirements.txt
setuptools
pelican~=4.9.1
Markdown~=3.4
checksumdir~=1.2.0

# Pelican plugins
beautifulsoup4~=4.12.3
# Dependencies of the vex plugin
jsonschema~=4.22.0
jsonref~=1.1.0

# Dev tools
invoke~=2.2.0
livereload~=2.6.3

# Patch for bug https://github.com/pallets/markupsafe/issues/284
markupsafe~=2.1.5

# VEX generation
cyclonedx-python-lib==11.5.0
packageurl-python==0.17.6
git+https://github.com/vex-generation-toolset/vex-generation-service.git@feat/make-python-package
PyGithub==2.8.1